Collapse des années dans l'historique, mode portrait dans phonecapture, et autres améliorations esthétiques.

+++ controlers/patient/actions/inc-action-saveOrdoForm.php
@@ -49,6 +49,19 @@ if (count($_POST)>2) {
         $supportID=$patient->createNewObjetByTypeName('ordoPorteur', '');
     }
 
+    // pour plus de clarté
+    if(isset($_POST['objetID']) and $supportID == $_POST['objetID']) {
+      $modeAction = 'edition';
+    } else {
+      $modeAction = 'renouv';
+    }
+
+    //par précaution on supprime le pdf antérieur
+    if($modeAction == 'edition') {
+      $doc= new msStockage();
+      $doc->setObjetID($supportID);
+      $doc->deleteDoc();
+    }
 
     //type d'impression modeprintObjetID
     if (isset($_POST['modeprintObjetID'])) {
@@ -64,11 +77,18 @@ if (count($_POST)>2) {
             } else {
                 $postObjetId='0';
             }
-            $id=$patient->createNewObjetByTypeName('ordoLigneOrdo', $v, $supportID, $m[1], $postObjetId);
+            if(!empty(trim($v))) {
+              $id=$patient->createNewObjetByTypeName('ordoLigneOrdo', $v, $supportID, $m[1], $postObjetId);
+            }
 
             if ($postObjetId>0) {
-                msSQL::sqlQuery("delete from objets_data where instance='".$postObjetId."' and typeID='".msData::getTypeIDFromName('ordoLigneOrdoALDouPas')."' ");
+                if($v=='' and $modeAction == 'edition') {
+                  $patient->setDeletedObjetAndSons($postObjetId);
+                } else {
+                  msSQL::sqlQuery("delete from objets_data where instance='".$postObjetId."' and typeID='".msData::getTypeIDFromName('ordoLigneOrdoALDouPas')."' ");
+                }
             }
+
             if (isset($_POST[$k.'CB'])) {
                 $patient->createNewObjetByTypeName('ordoLigneOrdoALDouPas', $_POST[$k.'CB'], $id);
             }
